Move xarray tests
